% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Makes a graph of the WorldPopulation data}
\usage{
CountryPopulation(country_name)
}
\arguments{
\item{country_name}{the name of a country in the data set}
}
\value{
a line graph of the given countries population from 1950 to 2020
}
\description{
Takes WorldPopulation data set and given a country name, returns a graph of
the population of that country from 1950 to 2020.
}
\examples{
CountryPopulation('United States of America')
CountryPopulation('China')
}
