% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{Each World Cup Series, the number of matches and attendance}
\format{
a data frame with 22 observations of 4 variables
\describe{
\item{WorldCup}{The Country and the year that the World Series was held}
\item{Matches}{The total matches the series had}
\item{Total_Attendance}{The total number of people who attended the series}
\item{Avg_Attendance}{The average attendance of each game in the series}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
At each World Cup, the total attendance, the average attendance of each game
and the number of matches each series had. Organized by the Country they were
held is as well as the year of each series.
}
\keyword{datasets}
