% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{Countries Population Each Year from 1950 to 2020}
\format{
a data frame with 16685 observations of 3 variables
\describe
{
\item{Country}{The country in the World}
\item{Year}{The year of the population}
\item{Population}{The population of that Country for that Year, in thousands}
}
}
\usage{
WorldPopulation
}
\description{
Countries Population Each Year from 1950 to 2020
}
\keyword{datasets}
